requirements.txt
datasets
langchain-groq
langgraph
python-dotenv
ragas